+++ instrumentation/cassandra/cassandra-4.0-common/testing/src/main/java/io/opentelemetry/cassandra/v4_0/common/AbstractCassandraTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.cassandra.v4.common;
+package io.opentelemetry.cassandra.v4_0.common;
 
 import static io.opentelemetry.instrumentation.testing.junit.db.SemconvStabilityUtil.maybeStable;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
