+++ instrumentation/couchbase/couchbase-2.0-common/testing/src/main/java/io/opentelemetry/instrumentation/couchbase/v2_0/common/AbstractCouchbaseAsyncClientTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.couchbase;
+package io.opentelemetry.instrumentation.couchbase.v2_0.common;
 
 import static io.opentelemetry.instrumentation.testing.junit.db.SemconvStabilityUtil.maybeStable;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
