+++ instrumentation/couchbase/couchbase-2.0-common/testing/src/main/java/io/opentelemetry/instrumentation/couchbase/v2_0/common/springdata/AbstractCouchbaseSpringRepositoryTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.couchbase.springdata;
+package io.opentelemetry.instrumentation.couchbase.v2_0.common.springdata;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,7 +13,7 @@ import com.couchbase.client.java.env.CouchbaseEnvironment;
 import com.couchbase.client.java.view.DefaultView;
 import com.couchbase.client.java.view.DesignDocument;
 import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.instrumentation.couchbase.AbstractCouchbaseTest;
+import io.opentelemetry.instrumentation.couchbase.v2_0.common.AbstractCouchbaseTest;
 import io.opentelemetry.instrumentation.testing.internal.AutoCleanupExtension;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
@@ -52,7 +52,7 @@ public abstract class AbstractCouchbaseSpringRepositoryTest extends AbstractCouc
                     DefaultView.create(
                         "all",
                         "function (doc, meta) {"
-                            + "  if (doc._class == \"io.opentelemetry.instrumentation.couchbase.springdata.TestDocument\") {"
+                            + "  if (doc._class == \"io.opentelemetry.instrumentation.couchbase.v2_0.common.springdata.TestDocument\") {"
                             + "    emit(meta.id, null);"
                             + " }"
                             + "}"))));
