+++ instrumentation/couchbase/couchbase-2.0-common/testing/src/main/java/io/opentelemetry/instrumentation/couchbase/v2_0/common/springdata/AbstractCouchbaseSpringTemplateTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.couchbase.springdata;
+package io.opentelemetry.instrumentation.couchbase.v2_0.common.springdata;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Named.named;
@@ -14,7 +14,7 @@ import com.couchbase.client.java.CouchbaseCluster;
 import com.couchbase.client.java.cluster.ClusterManager;
 import com.couchbase.client.java.env.CouchbaseEnvironment;
 import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.instrumentation.couchbase.AbstractCouchbaseTest;
+import io.opentelemetry.instrumentation.couchbase.v2_0.common.AbstractCouchbaseTest;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import java.util.ArrayList;
