+++ instrumentation/rxjava/rxjava-3.0-common/library/src/main/java/io/opentelemetry/instrumentation/rxjava/v3_0/common/RxJava3AsyncOperationEndStrategyBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.rxjava.v3.common;
+package io.opentelemetry.instrumentation.rxjava.v3_0.common;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
