+++ instrumentation/rxjava/rxjava-3.0-common/testing/src/main/java/io/opentelemetry/instrumentation/rxjava/v3_0/common/AbstractTracedWithSpan.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.rxjava.v3.common;
+package io.opentelemetry.instrumentation.rxjava.v3_0.common;
 
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Flowable;
