+++ instrumentation/rxjava/rxjava-3.0/javaagent/src/test/java/RxJava3Test.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import io.opentelemetry.instrumentation.rxjava.v3.common.AbstractRxJava3Test;
+import io.opentelemetry.instrumentation.rxjava.v3_0.common.AbstractRxJava3Test;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import org.junit.jupiter.api.extension.RegisterExtension;
