+++ instrumentation/rxjava/rxjava-3.0/library/src/test/java/RxJava3Test.java
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import io.opentelemetry.instrumentation.rxjava.v3.common.AbstractRxJava3Test;
 import io.opentelemetry.instrumentation.rxjava.v3_0.TracingAssembly;
+import io.opentelemetry.instrumentation.rxjava.v3_0.common.AbstractRxJava3Test;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.LibraryInstrumentationExtension;
 import org.junit.jupiter.api.BeforeAll;
